Added most of functionality for persistent history
COMMANDS method 'scrollback' with three modes: read, write, purge.

Read and write are fully implemented, purge is not yet. Ideally for
purge will expose options to mods and also take configuration from
json of how many days to consistently purge up to. Cane use JS's
native Date object for this.

In current version, everything -should- work though getting an error
where the call to 'this.scrollback' under 'join()' is saying that
'this.scrollback' is undefined rather than a function. Should be pretty
quick to find, will take a rest and look later.

Looking at config.json should be obvious what new options there are.
I've written the structure to accept storing logs in a datastore as
an alternative, but most of them just return blank or raise an error
for now as storing stringified JSON to a text file line-by-line is
super simple and pretty fast with Node's streaming api's.